services/visualization/requirements.txt
influxdb_client==1.37.0
paho_mqtt==1.6.1
python-dotenv==1.0.0
